# specs/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Information Systems Labs API
  version: 1.0.0
  description: |
    REST API for managing routes, vehicles, retail points, drivers, orders and related resources.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/vehicles:
    post:
      tags: [Vehicles]
      summary: Create a vehicle
      operationId: createVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveVehicle'
      responses:
        '201':
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Vehicles]
      summary: List vehicles
      operationId: listVehicles
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/VehicleSort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Paged list of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/vehicles/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Vehicles]
      summary: Get vehicle by id
      operationId: getVehicle
      responses:
        '200':
          description: Vehicle found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Vehicles]
      summary: Update vehicle
      operationId: updateVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveVehicle'
      responses:
        '200':
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Vehicles]
      summary: Delete vehicle
      operationId: deleteVehicle
      responses:
        '204':
          description: Vehicle removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/route-points:
    post:
      tags: [Route Points]
      summary: Create a route point
      operationId: createRoutePoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRoutePoint'
      responses:
        '201':
          description: Route point created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Route Points]
      summary: List route points
      operationId: listRoutePoints
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/RoutePointSort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Paged list of route points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageRoutePoint'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/route-points/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Route Points]
      summary: Get route point by id
      operationId: getRoutePoint
      responses:
        '200':
          description: Route point found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Route Points]
      summary: Update route point
      operationId: updateRoutePoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRoutePoint'
      responses:
        '200':
          description: Route point updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Route Points]
      summary: Delete route point
      operationId: deleteRoutePoint
      responses:
        '204':
          description: Route point removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/route-points/top-retail-points:
    parameters:
      - $ref: '#/components/parameters/Limit'
    get:
      tags: [Route Points]
      summary: Get top retail points by visit count
      operationId: getTopRetailPoints
      responses:
        '200':
          description: List of most visited retail points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetailPoint'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/routes:
    post:
      tags: [Routes]
      summary: Create a route
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRoute'
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Routes]
      summary: List routes
      operationId: listRoutes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/RouteSort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Paged list of routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageRoute'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/routes/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Routes]
      summary: Get route by id
      operationId: getRoute
      responses:
        '200':
          description: Route found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Routes]
      summary: Update route
      operationId: updateRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRoute'
      responses:
        '200':
          description: Route updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Routes]
      summary: Delete route
      operationId: deleteRoute
      responses:
        '204':
          description: Route removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/routes/within-period:
    get:
      tags: [Routes]
      summary: Get routes within period
      operationId: getRoutesWithinPeriod
      parameters:
        - $ref: '#/components/parameters/PeriodStart'
        - $ref: '#/components/parameters/PeriodEnd'
      responses:
        '200':
          description: Routes within period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Route'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/routes/retail-point/{retailPointId}:
    parameters:
      - name: retailPointId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Routes]
      summary: Get routes by retail point id
      operationId: getRoutesByRetailPoint
      responses:
        '200':
          description: Routes related to the retail point
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Route'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
...
    VehicleSort:
      name: sort
      in: query
      description: >
        Повторяемый параметр. Поля: id, gosNumber, tonnageInTons, bodyHeightInMeters,
        bodyWidthInMeters, bodyLengthInCubicMeters.
      schema:
        type: array
        items:
          type: string
          pattern: '^(id|gosNumber|tonnageInTons|bodyHeightInMeters|bodyWidthInMeters|bodyLengthInCubicMeters)(,(?i:asc|desc))?$'
      style: form
      explode: true
...
    RoutePointSort:
      name: sort
      in: query
      description: >
        Повторяемый параметр. Поля: id, orderNumber, plannedStartTime, plannedEndTime.
      schema:
        type: array
        items:
          type: string
          pattern: '^(id|orderNumber|plannedStartTime|plannedEndTime)(,(?i:asc|desc))?$'
      style: form
      explode: true
    RouteSort:
      name: sort
      in: query
      description: >
        Повторяемый параметр. Поля: id, creationTime, plannedStartTime, plannedEndTime, mileageInKm.
      schema:
        type: array
        items:
          type: string
          pattern: '^(id|creationTime|plannedStartTime|plannedEndTime|mileageInKm)(,(?i:asc|desc))?$'
      style: form
      explode: true
    Filter:
      name: filter
      in: query
      style: deepObject
      explode: true
      schema:
        type: object
        additionalProperties:
          type: string
    Limit:
      name: limit
      in: query
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
    PeriodStart:
      name: periodStart
      in: query
      required: true
      schema:
        type: string
        format: date-time
    PeriodEnd:
      name: periodEnd
      in: query
      required: true
      schema:
        type: string
        format: date-time
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema:
            type: string
  schemas:
    Driver:
      type: object
      required: [id, firstName, lastName, passport]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        firstName:
          type: string
          minLength: 1
        middleName:
          type: string
          nullable: true
        lastName:
          type: string
          minLength: 1
        passport:
          type: string
          minLength: 1
...
    Order:
      type: object
      required: [id, goodsType, minTemperature, maxTemperature, volumeInCubicMeters, weightInKg]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        goodsType:
          type: string
          minLength: 1
        minTemperature:
          type: integer
          format: int32
          minimum: -2147483648
          maximum: 2147483647
        maxTemperature:
          type: integer
          format: int32
          minimum: -2147483648
          maximum: 2147483647
        volumeInCubicMeters:
          type: number
          format: double
          minimum: 0.001
        weightInKg:
          type: number
          format: double
          minimum: 0.001
...
    Point:
      type: object
      required: [longitude, latitude]
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
    RetailPoint:
      type: object
      required: [id, name, address, location, type, timezone]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        location:
          $ref: '#/components/schemas/Point'
        type:
          $ref: '#/components/schemas/PointType'
        timezone:
          type: string
          minLength: 1
...
    Vehicle:
      type: object
      required: [id, driver, gosNumber, tonnageInTons, bodyHeightInMeters, bodyWidthInMeters, bodyLengthInCubicMeters]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        driver:
          $ref: '#/components/schemas/Driver'
        gosNumber:
          type: string
          minLength: 2
        tonnageInTons:
          type: number
          format: double
          minimum: 0.01
        bodyHeightInMeters:
          type: number
          format: double
          minimum: 0.01
        bodyWidthInMeters:
          type: number
          format: double
          minimum: 0.01
        bodyLengthInCubicMeters:
          type: number
          format: double
          minimum: 0.01
    SaveVehicle:
      type: object
      required: [driverId, gosNumber, tonnageInTons, bodyHeightInMeters, bodyWidthInMeters, bodyLengthInCubicMeters]
      properties:
        driverId:
          type: integer
          format: int64
          minimum: 1
        gosNumber:
          type: string
          minLength: 1
        tonnageInTons:
          type: number
          format: double
          minimum: 0.01
        bodyHeightInMeters:
          type: number
          format: double
          minimum: 0.01
        bodyWidthInMeters:
          type: number
          format: double
          minimum: 0.01
        bodyLengthInCubicMeters:
          type: number
          format: double
          minimum: 0.01
    Route:
      type: object
      required: [id, routePoints, vehicle, creationTime, plannedStartTime, plannedEndTime, mileageInKm]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        routePoints:
          type: array
          items:
            $ref: '#/components/schemas/RoutePoint'
        vehicle:
          $ref: '#/components/schemas/Vehicle'
        creationTime:
          type: string
          format: date-time
        plannedStartTime:
          type: string
          format: date-time
        plannedEndTime:
          type: string
          format: date-time
        mileageInKm:
          type: number
          format: double
          minimum: 0.001
    SaveRoute:
      type: object
      required: [routePoints, vehicleId, creationTime, plannedStartTime, plannedEndTime, mileageInKm]
      properties:
        routePoints:
          type: array
          items:
            $ref: '#/components/schemas/SaveRoutePoint'
            nullable: false
        vehicleId:
          type: integer
          format: int64
          minimum: 1
        creationTime:
          type: string
          format: date-time
        plannedStartTime:
          type: string
          format: date-time
        plannedEndTime:
          type: string
          format: date-time
        mileageInKm:
          type: number
          format: double
          minimum: 0.001
    RoutePoint:
      type: object
      required: [id, routeId, retailPoint, operationType, orders, plannedStartTime, plannedEndTime, orderNumber]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        routeId:
          type: integer
          format: int64
          minimum: 1
        retailPoint:
          $ref: '#/components/schemas/RetailPoint'
        operationType:
          $ref: '#/components/schemas/OperationType'
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        plannedStartTime:
          type: string
          format: date-time
        plannedEndTime:
          type: string
          format: date-time
        orderNumber:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
    SaveRoutePoint:
      type: object
      required: [retailPointId, operationType, orderIds, plannedStartTime, plannedEndTime, orderNumber]
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        routeId:
          type: integer
          format: int64
          nullable: false
          minimum: 1
        retailPointId:
          type: integer
          format: int64
          minimum: 1
        operationType:
          $ref: '#/components/schemas/OperationType'
        orderIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
        plannedStartTime:
          type: string
          format: date-time
        plannedEndTime:
          type: string
          format: date-time
        orderNumber:
          type: integer
          format: int32
...
    PageVehicle:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Vehicle'
...
    PageRoutePoint:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/RoutePoint'
    PageRoute:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Route'
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages, sort, first, last, empty]
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    OperationType:
      type: string
      enum: [LOAD, UNLOAD, VISIT]
    PointType:
      type: string
      enum: [SHOP, WAREHOUSE, GARAGE]